Fix s3 이미지 등록시 불필요하게 로컬에 저장하는 부분 제거

+++ src/main/java/com/example/eunboard/member/application/service/MemberService.java
@@ -114,9 +114,11 @@ public class MemberService implements MemberUseCase {
         Member member = memberRepository.findById(memberId).get();
         // 이미지 존재시 save
         if (multipartFile != null) {
-            // 기존 이미지 삭제
-            fileUploadUtils.delete(member.getProfileImage());
-            String upload = fileUploadUtils.upload(multipartFile, "image/profiles/" + memberId);
+            // 기존 이미지 삭제 하지만 기본 이미지는 삭제하면 안됨.
+            if (!member.getProfileImage().equals(fileUploadUtils.getDefaultImageUrl())) {
+                fileUploadUtils.delete(member.getProfileImage());
+            }
+            String upload = fileUploadUtils.upload(multipartFile, "mate/image/profiles/" + memberId);
             member.setProfileImage(upload);
             memberRepository.save(member);
         }
